+++ packages/bridge-controller/src/selectors.test.ts
@@ -1372,6 +1372,10 @@ describe('Bridge Selectors', () => {
         '1': {
           isActiveSrc: true,
           isActiveDest: true,
+          batchSellDestStablecoins: [
+            'eip155:1/erc20:0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48',
+            'eip155:1/slip44:60',
+          ],
         },
         '10': {
           isActiveSrc: true,
@@ -1424,6 +1428,10 @@ describe('Bridge Selectors', () => {
           'eip155:1': {
             isActiveSrc: true,
             isActiveDest: true,
+            batchSellDestStablecoins: [
+              'eip155:1/erc20:0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48',
+              'eip155:1/slip44:60',
+            ],
           },
           'eip155:10': {
             isActiveSrc: true,