+++ packages/controller-utils/src/create-service-policy.test.ts
@@ -86,7 +86,7 @@ describe('createServicePolicy', () => {
       });
 
       describe('if the service execution time is beyond the threshold', () => {
-        it('calls onDegraded listeners once', async () => {
+        it('calls onDegraded listeners once with the execution time', async () => {
           const delay = threshold + 1;
           const mockService = jest.fn(() => {
             return new Promise<void>((resolve) => {
@@ -102,6 +102,9 @@ describe('createServicePolicy', () => {
           await promise;
 
           expect(onDegradedListener).toHaveBeenCalledTimes(1);
+          expect(onDegradedListener).toHaveBeenCalledWith({
+            duration: delay,
+          });
         });
 
         it('does not call onAvailable listeners', async () => {
@@ -1473,7 +1476,7 @@ describe('createServicePolicy', () => {
           });
 
           describe('if the service execution time is beyond the threshold', () => {
-            it('calls onDegraded listeners once', async () => {
+            it('calls onDegraded listeners once with the execution time', async () => {
               let invocationCounter = 0;
               const delay = threshold + 1;
               const mockService = (): Promise<{ some: string }> => {
@@ -1498,6 +1501,9 @@ describe('createServicePolicy', () => {
               await promise;
 
               expect(onDegradedListener).toHaveBeenCalledTimes(1);
+              expect(onDegradedListener).toHaveBeenCalledWith({
+                duration: delay,
+              });
             });
 
             it('does not call onAvailable listeners', async () => {
@@ -1657,7 +1663,7 @@ describe('createServicePolicy', () => {
             });
 
             describe('if the service execution time is beyond the threshold', () => {
-              it('calls onDegraded listeners once', async () => {
+              it('calls onDegraded listeners once with the execution time', async () => {
                 const maxConsecutiveFailures = DEFAULT_MAX_RETRIES + 2;
                 const delay = threshold + 1;
                 let invocationCounter = 0;
@@ -1686,6 +1692,9 @@ describe('createServicePolicy', () => {
                 await promise;
 
                 expect(onDegradedListener).toHaveBeenCalledTimes(1);
+                expect(onDegradedListener).toHaveBeenCalledWith({
+                  duration: delay,
+                });
               });
 
               it('does not call onAvailable listeners', async () => {
@@ -1851,7 +1860,7 @@ describe('createServicePolicy', () => {
             });
 
             describe('if the service execution time is beyond the threshold', () => {
-              it('calls onDegraded listeners once', async () => {
+              it('calls onDegraded listeners once with the execution time', async () => {
                 const maxConsecutiveFailures = DEFAULT_MAX_RETRIES + 1;
                 const delay = threshold + 1;
                 let invocationCounter = 0;
@@ -1880,6 +1889,9 @@ describe('createServicePolicy', () => {
                 await promise;
 
                 expect(onDegradedListener).toHaveBeenCalledTimes(1);
+                expect(onDegradedListener).toHaveBeenCalledWith({
+                  duration: delay,
+                });
               });
 
               it('does not call onAvailable listeners', async () => {
@@ -2258,7 +2270,7 @@ describe('createServicePolicy', () => {
             });
 
             describe('if the service execution time is beyond the threshold', () => {
-              it('calls onDegraded listeners once', async () => {
+              it('calls onDegraded listeners once with the execution time', async () => {
                 const maxRetries = DEFAULT_MAX_CONSECUTIVE_FAILURES - 2;
                 const delay = threshold + 1;
                 let invocationCounter = 0;
@@ -2284,6 +2296,9 @@ describe('createServicePolicy', () => {
                 await promise;
 
                 expect(onDegradedListener).toHaveBeenCalledTimes(1);
+                expect(onDegradedListener).toHaveBeenCalledWith({
+                  duration: delay,
+                });
               });
 
               it('does not call onAvailable listeners', async () => {
@@ -2432,7 +2447,7 @@ describe('createServicePolicy', () => {
             });
 
             describe('if the service execution time is beyond the threshold', () => {
-              it('calls onDegraded listeners once', async () => {
+              it('calls onDegraded listeners once with the execution time', async () => {
                 const maxRetries = DEFAULT_MAX_CONSECUTIVE_FAILURES - 1;
                 const delay = DEFAULT_DEGRADED_THRESHOLD + 1;
                 let invocationCounter = 0;
@@ -2458,6 +2473,9 @@ describe('createServicePolicy', () => {
                 await promise;
 
                 expect(onDegradedListener).toHaveBeenCalledTimes(1);
+                expect(onDegradedListener).toHaveBeenCalledWith({
+                  duration: delay,
+                });
               });
 
               it('does not call onAvailable listeners', async () => {
@@ -2807,7 +2825,7 @@ describe('createServicePolicy', () => {
             });
 
             describe('if the service execution time is beyond the threshold', () => {
-              it('calls onDegraded listeners once', async () => {
+              it('calls onDegraded listeners once with the execution time', async () => {
                 const maxConsecutiveFailures = 5;
                 const maxRetries = maxConsecutiveFailures - 2;
                 const delay = DEFAULT_DEGRADED_THRESHOLD + 1;
@@ -2838,6 +2856,9 @@ describe('createServicePolicy', () => {
                 await promise;
 
                 expect(onDegradedListener).toHaveBeenCalledTimes(1);
+                expect(onDegradedListener).toHaveBeenCalledWith({
+                  duration: delay,
+                });
               });
 
               it('does not call onAvailable listeners', async () => {
@@ -3014,7 +3035,7 @@ describe('createServicePolicy', () => {
             });
 
             describe('if the service execution time is beyond the threshold', () => {
-              it('calls onDegraded listeners once', async () => {
+              it('calls onDegraded listeners once with the execution time', async () => {
                 const maxConsecutiveFailures = 5;
                 const maxRetries = maxConsecutiveFailures - 1;
                 const delay = threshold + 1;
@@ -3045,6 +3066,9 @@ describe('createServicePolicy', () => {
                 await promise;
 
                 expect(onDegradedListener).toHaveBeenCalledTimes(1);
+                expect(onDegradedListener).toHaveBeenCalledWith({
+                  duration: delay,
+                });
               });
 
               it('does not call onAvailable listeners', async () => {